main関数の分割
