+++ src/main.ts
@@ -1,9 +1,7 @@
 import * as THREE from "three";
-import { GLTFLoader } from "three/addons/loaders/GLTFLoader.js";
 import { OrbitControls } from "three/addons/controls/OrbitControls.js";
-import { frameArea } from "./ref/frameArea.ts";
 import { resizeRendererToDisplaySize } from "./ref/resizeRender.ts";
-import { objectSetting } from "./ref/objectSetting.ts";
+import { Road } from "./ref/ground/road.ts";
 
 function main() {
   const canvas = document.querySelector("#c") as HTMLCanvasElement;
@@ -29,26 +27,6 @@ function main() {
     const light = new THREE.HemisphereLight(skyColor, groundColor, intensity);
     scene.add(light);
   }
-
-  {
-    objectSetting(scene);
-  }
-
-  {
-    const gltfLoader = new GLTFLoader();
-    const url = "src/3d/model/KC3.gltf";
-    gltfLoader.load(url, (gltf) => {
-      const root = gltf.scene;
-      scene.add(root);
-      const box = new THREE.Box3().setFromObject(root);
-      const boxSize = box.getSize(new THREE.Vector3()).length();
-      const boxCenter = box.getCenter(new THREE.Vector3());
-      frameArea(boxSize * 0.5, boxSize, boxCenter, camera);
-      controls.maxDistance = boxSize * 10;
-      controls.target.copy(boxCenter);
-      controls.update();
-    });
-  }
   function render() {
     if (resizeRendererToDisplaySize(renderer)) {
       const canvas = renderer.domElement;
@@ -65,3 +43,4 @@ function main() {
 }
 
 main();
+Road();
