+++ src/ref/frameArea.ts
@@ -1,6 +1,5 @@
 import * as THREE from "three";
 import { PerspectiveCamera } from "three";
-
 export function frameArea(
   sizeToFitOnScreen: number,
   boxSize: number,
@@ -16,18 +15,12 @@ export function frameArea(
     .subVectors(camera.position, boxCenter)
     .multiply(new THREE.Vector3(1, 2, 1)) //カメラの角度変更
     .normalize();
-
   // move the camera to a position distance units way from the center
   // in whatever direction the camera was from the center already
   camera.position.copy(direction.multiplyScalar(distance).add(boxCenter));
-
-  // pick some near and far values for the frustum that
-  // will contain the box.
   camera.near = boxSize / 100;
   camera.far = boxSize * 100;
-
   camera.updateProjectionMatrix();
-
   // point the camera to look at the center of the box
   camera.lookAt(boxCenter.x, boxCenter.y, boxCenter.z);
 }
